Update GLIF3.py such that the evolution is accurate regarding exactly when spike begins and ends
In this updated glif3, the dt has to be greater than the smallest spike window length. With this being hold true, the first order Euler forward type evolution of V and asc is only carried out at the part of each time bin excluding the end of the spike window from previously detected spikes, and excluding the head of the newly detected spike windows which overlaps with the current time bin.